Decouple the "Stats" component into a separate component called FilteredStats component for detailed statistical analysis and visualization
